requirements.txt
flask
pymongo
bson
requests
dialogflow
